refactor: move code so that pytest and pylint can both find it.
