--- test_utils/__init__.py
@@ -1,7 +1,0 @@
-"""
-Test utilities.
-
-Since pytest discourages putting __init__.py into test directory (i.e. making tests a package)
-one cannot import from anywhere under tests folder. However, some utility classes/methods might be useful
-in multiple test modules (i.e. factoryboy factories, base test classes). So this package is the place to put them.
-"""
