+++ tests/__init__.py
@@ -1,0 +1,3 @@
+"""
+Tests that the cookiecutters work properly, and helpers for the tests.
+"""
